+++ lib/query/completer_readline_test.go
@@ -280,6 +280,7 @@ var completerStatementsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1074,6 +1075,7 @@ var completerWithArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1147,6 +1149,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1175,6 +1178,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1205,6 +1209,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1236,6 +1241,7 @@ var completerSelectArgsTests = []completerTest{
 		OrigLine: "select 1 into @var1 ",
 		Index:    20,
 		Expect: readline.CandidateList{
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1275,6 +1281,7 @@ var completerSelectArgsTests = []completerTest{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1301,6 +1308,7 @@ var completerSelectArgsTests = []completerTest{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1347,6 +1355,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1372,6 +1381,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1404,6 +1414,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1460,6 +1471,7 @@ var completerSelectArgsTests = []completerTest{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1529,6 +1541,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("CROSS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FULL"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -1569,6 +1582,7 @@ var completerSelectArgsTests = []completerTest{
 		Index:    25,
 		Expect: readline.CandidateList{
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
 			{Name: []rune("HAVING"), AppendSpace: true},
@@ -1606,6 +1620,7 @@ var completerSelectArgsTests = []completerTest{
 		Index:    36,
 		Expect: readline.CandidateList{
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("HAVING"), AppendSpace: true},
 			{Name: []rune("INTERSECT"), AppendSpace: true},
@@ -1629,6 +1644,7 @@ var completerSelectArgsTests = []completerTest{
 		Index:    26,
 		Expect: readline.CandidateList{
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("INTERSECT"), AppendSpace: true},
 			{Name: []rune("LIMIT"), AppendSpace: true},
@@ -1664,6 +1680,7 @@ var completerSelectArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("ASC")},
 			{Name: []rune("DESC")},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("LIMIT"), AppendSpace: true},
 			{Name: []rune("NULLS FIRST")},
@@ -1677,6 +1694,7 @@ var completerSelectArgsTests = []completerTest{
 		OrigLine: "select 1 from t1 where 1 order by 1 desc ",
 		Index:    41,
 		Expect: readline.CandidateList{
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("LIMIT"), AppendSpace: true},
 			{Name: []rune("NULLS FIRST")},
@@ -1700,6 +1718,7 @@ var completerSelectArgsTests = []completerTest{
 		OrigLine: "select 1 from t1 where 1 order by 1 desc nulls first ",
 		Index:    53,
 		Expect: readline.CandidateList{
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("LIMIT"), AppendSpace: true},
 			{Name: []rune("OFFSET"), AppendSpace: true},
@@ -1768,10 +1787,100 @@ var completerSelectArgsTests = []completerTest{
 		},
 	},
 	{
-		Name:     "SelectArgs After OFFSET Clause",
+		Name:     "SelectArgs After Value in OFFSET Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 2 ",
+		Index:    34,
+		Expect: readline.CandidateList{
+			{Name: []rune("FETCH"), AppendSpace: true},
+			{Name: []rune("FOR UPDATE")},
+			{Name: []rune("ROWS")},
+		},
+	},
+	{
+		Name:     "SelectArgs After Value in OFFSET Clause with LIMIT Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 limit 1 offset 2 ",
+		Index:    42,
+		Expect: readline.CandidateList{
+			{Name: []rune("FOR UPDATE")},
+			{Name: []rune("ROWS")},
+		},
+	},
+	{
+		Name:     "SelectArgs After 1 in OFFSET Clause",
 		Line:     "",
 		OrigLine: "select 1 from t1 where 1 offset 1 ",
 		Index:    34,
+		Expect: readline.CandidateList{
+			{Name: []rune("FETCH"), AppendSpace: true},
+			{Name: []rune("FOR UPDATE")},
+			{Name: []rune("ROW")},
+		},
+	},
+	{
+		Name:     "SelectArgs After FETCH without OFFSET Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 fetch ",
+		Index:    31,
+		Expect: readline.CandidateList{
+			{Name: []rune("FIRST"), AppendSpace: true},
+		},
+	},
+	{
+		Name:     "SelectArgs After FETCH",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 10 rows fetch ",
+		Index:    46,
+		Expect: readline.CandidateList{
+			{Name: []rune("NEXT"), AppendSpace: true},
+		},
+	},
+	{
+		Name:     "SelectArgs After Value in FETCH Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 10 rows fetch next 2 ",
+		Index:    53,
+		Expect: readline.CandidateList{
+			{Name: []rune("PERCENT"), AppendSpace: true},
+			{Name: []rune("ROWS"), AppendSpace: true},
+		},
+	},
+	{
+		Name:     "SelectArgs After 1 in FETCH Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 10 rows fetch next 1 ",
+		Index:    53,
+		Expect: readline.CandidateList{
+			{Name: []rune("PERCENT"), AppendSpace: true},
+			{Name: []rune("ROW"), AppendSpace: true},
+		},
+	},
+	{
+		Name:     "SelectArgs After ROW in FETCH Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 10 rows fetch next 1 row ",
+		Index:    57,
+		Expect: readline.CandidateList{
+			{Name: []rune("FOR UPDATE")},
+			{Name: []rune("ONLY")},
+			{Name: []rune("WITH TIES")},
+		},
+	},
+	{
+		Name:     "SelectArgs After WITH in FETCH Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 10 rows fetch next 1 row with ",
+		Index:    62,
+		Expect: readline.CandidateList{
+			{Name: []rune("TIES")},
+		},
+	},
+	{
+		Name:     "SelectArgs After ONLY in FETCH Clause",
+		Line:     "",
+		OrigLine: "select 1 from t1 where 1 offset 10 rows fetch next 1 row only ",
+		Index:    62,
 		Expect: readline.CandidateList{
 			{Name: []rune("FOR UPDATE")},
 		},
@@ -1885,6 +1994,7 @@ var completerInsertArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -2068,6 +2178,7 @@ var completerReplaceArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -2226,6 +2337,7 @@ var completerCreateArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -2246,6 +2358,7 @@ var completerCreateArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},
@@ -2638,6 +2751,7 @@ var completerDeclareArgsTests = []completerTest{
 		Expect: readline.CandidateList{
 			{Name: []rune("AS"), AppendSpace: true},
 			{Name: []rune("EXCEPT"), AppendSpace: true},
+			{Name: []rune("FETCH"), AppendSpace: true},
 			{Name: []rune("FOR UPDATE")},
 			{Name: []rune("FROM"), AppendSpace: true},
 			{Name: []rune("GROUP BY"), AppendSpace: true},